Move APIStatusBar into its own component
