+++ lib/hooks/use-api-version.ts
@@ -2,6 +2,7 @@ import useSWR from 'swr'
 
 import {ResponseError} from 'lib/utils/safe-fetch'
 
+const REFRESH_INTERVAL_MS = 10_000
 const VERSION_URL = process.env.NEXT_PUBLIC_API_URL + '/version'
 
 type Version = {
@@ -12,7 +13,8 @@ type Version = {
 
 export default function useApiVersion() {
   return useSWR<Version, ResponseError>(VERSION_URL, {
-    refreshInterval: 30_000, // 30 seconds
+    refreshInterval: REFRESH_INTERVAL_MS,
+    refreshWhenOffline: true,
     revalidateOnFocus: true
   })
 }
